Function vga::panic::rust_begin_unwind

#[lang = "panic_fmt"]
#[no_mangle]

pub extern "C" fn rust_begin_unwind(
    args: Arguments, 
    file: &'static str, 
    line: usize
) -> !

Called to handle a panic.

Since kernel panics are non-recoverable, this function prints out the error message and hangs forever.

Eventually – way in the future – when we have disk I/O and stuff, we'll probably want to write out some core dumps here as well.