Struct vga::Terminal

pub struct Terminal { /* fields omitted */ }

Methods

impl Terminal

fn x_position(&self) -> usize

const unsafe fn new(colors: Palette, buffer_start: usize) -> Terminal

Constructs a new Terminal for abuffer starting at the given address.

Arguments

colors: the default color palette for the terminal

buffer_start: the address of the to the memory location where the terminal's VGA buffer begins

Unsafe due to:

Casting a raw address to an array

fn set_colors(&mut self, bg: Color, fg: Color) -> &mut Self

Set the color palette used for writing subsequent characters.

fn clear(&mut self) -> &mut Self

Clear the terminal

fn write_byte(&mut self, byte: u8) -> &mut Self

Write the given byte to the terminal, and advance the cursor position.